All uncollectedfees endpoint
