+++ v3data/hypervisor.py
@@ -251,116 +251,6 @@ class HypervisorData:
         self.basics_data = basics
         self.pools_data = pools
 
-    async def _get_uncollected_fees_data(self, hypervisor_address):
-        hypervisor_query = """
-        query hypervisor($id: String!){
-            uniswapV3Hypervisor(
-                id: $id
-            ){
-                id
-                pool{
-                    id
-                    token0 {decimals}
-                    token1 {decimals}
-                }
-                baseLiquidity
-                baseLower
-                baseUpper
-                baseTokensOwed0
-                baseTokensOwed1
-                baseFeeGrowthInside0LastX128
-                baseFeeGrowthInside1LastX128
-                limitLiquidity
-                limitLower
-                limitUpper
-                limitTokensOwed0
-                limitTokensOwed1
-                limitFeeGrowthInside0LastX128
-                limitFeeGrowthInside1LastX128
-                conversion {
-                    baseTokenIndex
-                    priceTokenInBase
-                    priceBaseInUSD
-                }
-                tvlUSD
-            }
-        }
-        """
-        hypervisor_variables = {"id": hypervisor_address.lower()}
-        hypervisor_response = await self.gamma_client.query(
-            hypervisor_query, hypervisor_variables
-        )
-        hypervisor_data = hypervisor_response["data"]["uniswapV3Hypervisor"]
-
-        pool_query = """
-        query pool(
-            $poolAddress: String!
-            $baseLower: Int!
-            $baseUpper: Int!
-            $limitLower: Int!
-            $limitUpper: Int!
-        ){
-            pool(id: $poolAddress){
-                tick
-                feeGrowthGlobal0X128
-                feeGrowthGlobal1X128
-            }
-            baseLower: ticks(
-                where: {
-                poolAddress: $poolAddress
-                tickIdx: $baseLower
-                }
-            ){
-                tickIdx
-                feeGrowthOutside0X128
-                feeGrowthOutside1X128
-            },
-            baseUpper: ticks(
-                where: {
-                poolAddress: $poolAddress
-                tickIdx: $baseUpper
-                }
-            ){
-                tickIdx
-                feeGrowthOutside0X128
-                feeGrowthOutside1X128
-            },
-            limitLower: ticks(
-                where: {
-                poolAddress: $poolAddress
-                tickIdx: $limitLower
-                }
-            ){
-                tickIdx
-                feeGrowthOutside0X128
-                feeGrowthOutside1X128
-            },
-            limitUpper: ticks(
-                where: {
-                poolAddress: $poolAddress
-                tickIdx: $limitUpper
-                }
-            ){
-                tickIdx
-                feeGrowthOutside0X128
-                feeGrowthOutside1X128
-            }
-        }
-        """
-        variables = {
-            "poolAddress": hypervisor_data["pool"]["id"],
-            "baseLower": hypervisor_data["baseLower"],
-            "baseUpper": hypervisor_data["baseUpper"],
-            "limitLower": hypervisor_data["limitLower"],
-            "limitUpper": hypervisor_data["limitUpper"],
-        }
-        pool_response = await self.uniswap_client.query(pool_query, variables)
-        tick_data = pool_response["data"]
-
-        self.fees_data = {"hypervisor_data": hypervisor_data, "tick_data": tick_data}
-
-        return True
-
 
 class HypervisorInfo(HypervisorData):
     def empty_returns(self):
@@ -591,245 +481,3 @@ class HypervisorInfo(HypervisorData):
             returns["weekly"] = returns["daily"]
 
         return returns
-
-
-class UncollectedFees(HypervisorData):
-    async def output(self, hypervisor_address, get_data=True):
-        if get_data:
-            await self._get_uncollected_fees_data(hypervisor_address)
-
-        data = self.fees_data
-
-        decimals_0 = int(data["hypervisor_data"]["pool"]["token0"]["decimals"])
-        decimals_1 = int(data["hypervisor_data"]["pool"]["token1"]["decimals"])
-
-        try:
-            base_fees_0, base_fees_1 = self.calc_fees(
-                fee_growth_global_0=int(
-                    data["tick_data"]["pool"]["feeGrowthGlobal0X128"]
-                ),
-                fee_growth_global_1=int(
-                    data["tick_data"]["pool"]["feeGrowthGlobal1X128"]
-                ),
-                tick_current=int(data["tick_data"]["pool"]["tick"]),
-                tick_lower=int(data["hypervisor_data"]["baseLower"]),
-                tick_upper=int(data["hypervisor_data"]["baseUpper"]),
-                fee_growth_outside_0_lower=int(
-                    data["tick_data"]["baseLower"][0]["feeGrowthOutside0X128"]
-                ),
-                fee_growth_outside_1_lower=int(
-                    data["tick_data"]["baseLower"][0]["feeGrowthOutside1X128"]
-                ),
-                fee_growth_outside_0_upper=int(
-                    data["tick_data"]["baseUpper"][0]["feeGrowthOutside0X128"]
-                ),
-                fee_growth_outside_1_upper=int(
-                    data["tick_data"]["baseUpper"][0]["feeGrowthOutside1X128"]
-                ),
-                liquidity=int(data["hypervisor_data"]["baseLiquidity"]),
-                fee_growth_inside_last_0=int(
-                    data["hypervisor_data"]["baseFeeGrowthInside0LastX128"]
-                ),
-                fee_growth_inside_last_1=int(
-                    data["hypervisor_data"]["baseFeeGrowthInside1LastX128"]
-                ),
-            )
-        except IndexError:
-            base_fees_0 = 0
-            base_fees_1 = 0
-
-        base_tokens_owed_0 = float(data["hypervisor_data"]["baseTokensOwed0"])
-        base_tokens_owed_1 = float(data["hypervisor_data"]["baseTokensOwed1"])
-
-        try:
-            limit_fees_0, limit_fees_1 = self.calc_fees(
-                fee_growth_global_0=int(
-                    data["tick_data"]["pool"]["feeGrowthGlobal0X128"]
-                ),
-                fee_growth_global_1=int(
-                    data["tick_data"]["pool"]["feeGrowthGlobal1X128"]
-                ),
-                tick_current=int(data["tick_data"]["pool"]["tick"]),
-                tick_lower=int(data["hypervisor_data"]["limitLower"]),
-                tick_upper=int(data["hypervisor_data"]["limitUpper"]),
-                fee_growth_outside_0_lower=int(
-                    data["tick_data"]["limitLower"][0]["feeGrowthOutside0X128"]
-                ),
-                fee_growth_outside_1_lower=int(
-                    data["tick_data"]["limitLower"][0]["feeGrowthOutside1X128"]
-                ),
-                fee_growth_outside_0_upper=int(
-                    data["tick_data"]["limitUpper"][0]["feeGrowthOutside0X128"]
-                ),
-                fee_growth_outside_1_upper=int(
-                    data["tick_data"]["limitUpper"][0]["feeGrowthOutside1X128"]
-                ),
-                liquidity=int(data["hypervisor_data"]["limitLiquidity"]),
-                fee_growth_inside_last_0=int(
-                    data["hypervisor_data"]["limitFeeGrowthInside0LastX128"]
-                ),
-                fee_growth_inside_last_1=int(
-                    data["hypervisor_data"]["limitFeeGrowthInside1LastX128"]
-                ),
-            )
-        except IndexError:
-            limit_fees_0 = 0
-            limit_fees_1 = 0
-
-        limit_tokens_owed_0 = float(data["hypervisor_data"]["limitTokensOwed0"])
-        limit_tokens_owed_1 = float(data["hypervisor_data"]["limitTokensOwed1"])
-
-        # Convert to USD
-        baseTokenIndex = int(data["hypervisor_data"]["conversion"]["baseTokenIndex"])
-        priceTokenInBase = float(
-            data["hypervisor_data"]["conversion"]["priceTokenInBase"]
-        )
-        priceBaseInUSD = float(data["hypervisor_data"]["conversion"]["priceBaseInUSD"])
-
-        if baseTokenIndex == 0:
-            base_fees_0_usd = base_fees_0 * priceBaseInUSD
-            base_fees_1_usd = base_fees_1 * priceTokenInBase * priceBaseInUSD
-            base_tokens_owed_0_usd = base_tokens_owed_0 * priceBaseInUSD
-            base_tokens_owed_1_usd = (
-                base_tokens_owed_1 * priceTokenInBase * priceBaseInUSD
-            )
-            limit_fees_0_usd = limit_fees_0 * priceBaseInUSD
-            limit_fees_1_usd = limit_fees_1 * priceTokenInBase * priceBaseInUSD
-            limit_tokens_owed_0_usd = limit_tokens_owed_0 * priceBaseInUSD
-            limit_tokens_owed_1_usd = (
-                limit_tokens_owed_1 * priceTokenInBase * priceBaseInUSD
-            )
-        elif baseTokenIndex == 1:
-            base_fees_0_usd = base_fees_0 * priceTokenInBase * priceBaseInUSD
-            base_fees_1_usd = base_fees_1 * priceBaseInUSD
-            base_tokens_owed_0_usd = (
-                base_tokens_owed_0 * priceTokenInBase * priceBaseInUSD
-            )
-            base_tokens_owed_1_usd = base_tokens_owed_1 * priceBaseInUSD
-            limit_fees_0_usd = limit_fees_0 * priceTokenInBase * priceBaseInUSD
-            limit_fees_1_usd = limit_fees_1 * priceBaseInUSD
-            limit_tokens_owed_0_usd = (
-                limit_tokens_owed_0 * priceTokenInBase * priceBaseInUSD
-            )
-            limit_tokens_owed_1_usd = limit_tokens_owed_1 * priceBaseInUSD
-        else:
-            base_fees_0_usd = 0
-            base_fees_1_usd = 0
-            base_tokens_owed_0_usd = 0
-            base_tokens_owed_1_usd = 1
-            limit_fees_0_usd = 0
-            limit_fees_1_usd = 0
-            limit_tokens_owed_0_usd = 0
-            limit_tokens_owed_1_usd = 0
-
-        uncollected_fees_total = (
-            base_fees_0_usd
-            + base_fees_1_usd
-            + base_tokens_owed_0_usd
-            + base_tokens_owed_1_usd
-            + limit_fees_0_usd
-            + limit_fees_1_usd
-            + limit_tokens_owed_0_usd
-            + limit_tokens_owed_1_usd
-        )
-
-        return {
-            "baseFees0": base_fees_0 / 10**decimals_0,
-            "baseFees1": base_fees_1 / 10**decimals_1,
-            "baseTokensOwed0": base_tokens_owed_0 / 10**decimals_0,
-            "baseTokensOwed1": base_tokens_owed_1 / 10**decimals_1,
-            "limitFees0": limit_fees_0 / 10**decimals_0,
-            "limitFees1": limit_fees_1 / 10**decimals_1,
-            "limitTokensOwed0": limit_tokens_owed_0 / 10**decimals_0,
-            "limitTokensOwed1": limit_tokens_owed_1 / 10**decimals_1,
-            "baseFees0USD": base_fees_0_usd,
-            "baseFees1USD": base_fees_1_usd,
-            "baseTokensOwed0USD": base_tokens_owed_0_usd,
-            "baseTokensOwed1USD": base_tokens_owed_1_usd,
-            "limitFees0USD": limit_fees_0_usd,
-            "limitFees1USD": limit_fees_1_usd,
-            "limitTokensOwed0USD": limit_tokens_owed_0_usd,
-            "limitTokensOwed1USD": limit_tokens_owed_1_usd,
-            "totalFees0": (base_fees_0 + base_tokens_owed_0 + limit_fees_0 + limit_tokens_owed_0) / 10**decimals_0,
-            "totalFees1": (base_fees_1 + base_tokens_owed_1 + limit_fees_1 + limit_tokens_owed_1) / 10**decimals_1,
-            "totalFeesUSD": uncollected_fees_total,
-        }
-
-    async def output_for_returns_calc(self, hypervisor_address, get_data=True):
-        fees = await self.output(hypervisor_address, get_data)
-
-        gross_fees = max(
-            (
-                fees["base_fees_0_usd"]
-                + fees["base_fees_1_usd"]
-                + fees["limit_fees_0_usd"]
-                + fees["limit_fees_1_usd"]
-            ),
-            0,
-        )
-
-        return {
-            "id": "uncollected_fees",
-            "timestamp": timestamp_ago(timedelta(0)),
-            "grossFeesUSD": gross_fees,
-            "protocolFeesUSD": gross_fees * 0.1,
-            "netFeesUSD": gross_fees * 0.9,
-            "totalAmountUSD": fees["tvl_usd"],
-        }
-
-    @staticmethod
-    def calc_fees(
-        fee_growth_global_0,
-        fee_growth_global_1,
-        tick_current,
-        tick_lower,
-        tick_upper,
-        fee_growth_outside_0_lower,
-        fee_growth_outside_1_lower,
-        fee_growth_outside_0_upper,
-        fee_growth_outside_1_upper,
-        liquidity,
-        fee_growth_inside_last_0,
-        fee_growth_inside_last_1,
-    ):
-        X128 = 2**128
-
-        if tick_current >= tick_lower:
-            fee_growth_below_pos_0 = fee_growth_outside_0_lower
-            fee_growth_below_pos_1 = fee_growth_outside_1_lower
-        else:
-            fee_growth_below_pos_0 = sub_in_256(
-                fee_growth_global_0, fee_growth_outside_0_lower
-            )
-            fee_growth_below_pos_1 = sub_in_256(
-                fee_growth_global_1, fee_growth_outside_1_lower
-            )
-
-        if tick_current >= tick_upper:
-            fee_growth_above_pos_0 = sub_in_256(
-                fee_growth_global_0, fee_growth_outside_0_upper
-            )
-            fee_growth_above_pos_1 = sub_in_256(
-                fee_growth_global_1, fee_growth_outside_1_upper
-            )
-        else:
-            fee_growth_above_pos_0 = fee_growth_outside_0_upper
-            fee_growth_above_pos_1 = fee_growth_outside_1_upper
-
-        fees_accum_now_0 = sub_in_256(
-            sub_in_256(fee_growth_global_0, fee_growth_below_pos_0),
-            fee_growth_above_pos_0,
-        )
-        fees_accum_now_1 = sub_in_256(
-            sub_in_256(fee_growth_global_1, fee_growth_below_pos_1),
-            fee_growth_above_pos_1,
-        )
-
-        uncollectedFees_0 = (
-            liquidity * (sub_in_256(fees_accum_now_0, fee_growth_inside_last_0))
-        ) / X128
-        uncollectedFees_1 = (
-            liquidity * (sub_in_256(fees_accum_now_1, fee_growth_inside_last_1))
-        ) / X128
-
-        return uncollectedFees_0, uncollectedFees_1
