+++ v3data/routers/celo.py
@@ -66,6 +66,13 @@ async def hypervisor_apy(response: Response, hypervisor_address):
     )
 
 
+@router.get("/hypervisor/{hypervisor_address}/uncollectedFees")
+async def hypervisor_uncollected_fees(hypervisor_address: str):
+    return await v3data.common.hypervisor.uncollected_fees(
+        CHAIN_CELO, hypervisor_address
+    )
+
+
 @router.get("/hypervisors/aggregateStats")
 async def aggregate_stats():
     return await v3data.common.hypervisor.aggregate_stats(CHAIN_CELO)
@@ -82,6 +89,11 @@ async def hypervisors_all():
     return await v3data.common.hypervisor.hypervisors_all(CHAIN_CELO)
 
 
+@router.get("/hypervisors/uncollectedFees")
+async def uncollected_fees_all():
+    return await v3data.common.hypervisor.uncollected_fees_all(CHAIN_CELO)
+
+
 @router.get("/user/{address}")
 async def user_data(address: str):
     return await v3data.common.users.user_data(CHAIN_CELO, address)
@@ -90,10 +102,3 @@ async def user_data(address: str):
 @router.get("/vault/{address}")
 async def account_data(address: str):
     return await v3data.common.users.account_data(CHAIN_CELO, address)
-
-
-@router.get("/hypervisor/{hypervisor_address}/uncollectedFees")
-async def hypervisor_uncollected_fees(hypervisor_address: str):
-    return await v3data.common.hypervisor.uncollected_fees(
-        CHAIN_CELO, hypervisor_address
-    )
