+++ v3data/routers/optimism.py
@@ -41,6 +41,7 @@ async def base_range_chart(hypervisor_address: str, days: int = 20):
         CHAIN_OPTIMISM, hypervisor_address, days
     )
 
+
 @router.get("/charts/benchmark/{hypervisor_address}")
 # @cache(expire=CHARTS_CACHE_TIMEOUT)
 async def benchmark_chart(
@@ -49,6 +50,7 @@ async def benchmark_chart(
     return await v3data.common.charts.benchmark_chart(
         CHAIN_OPTIMISM, hypervisor_address, startDate, endDate
     )
+
 
 @router.get("/hypervisor/{hypervisor_address}/basicStats")
 async def hypervisor_basic_stats(hypervisor_address, response: Response):
@@ -62,6 +64,13 @@ async def hypervisor_basic_stats(hypervisor_address, response: Response):
 async def hypervisor_apy(response: Response, hypervisor_address):
     return await v3data.common.hypervisor.hypervisor_apy(
         CHAIN_OPTIMISM, hypervisor_address, response
+    )
+
+
+@router.get("/hypervisor/{hypervisor_address}/uncollectedFees")
+async def hypervisor_uncollected_fees(hypervisor_address: str):
+    return await v3data.common.hypervisor.uncollected_fees(
+        CHAIN_OPTIMISM, hypervisor_address
     )
 
 
@@ -79,6 +88,11 @@ async def hypervisors_return():
 @router.get("/hypervisors/allData")
 async def hypervisors_all():
     return await v3data.common.hypervisor.hypervisors_all(CHAIN_OPTIMISM)
+
+
+@router.get("/hypervisors/uncollectedFees")
+async def uncollected_fees_all():
+    return await v3data.common.hypervisor.uncollected_fees_all(CHAIN_OPTIMISM)
 
 
 @router.get("/allRewards")
@@ -99,10 +113,3 @@ async def user_data(address: str):
 @router.get("/vault/{address}")
 async def account_data(address: str):
     return await v3data.common.users.account_data(CHAIN_OPTIMISM, address)
-
-
-@router.get("/hypervisor/{hypervisor_address}/uncollectedFees")
-async def hypervisor_uncollected_fees(hypervisor_address: str):
-    return await v3data.common.hypervisor.uncollected_fees(
-        CHAIN_OPTIMISM, hypervisor_address
-    )
